Fixes #154, Correctly build the node tree from JS
Attempting to use more than one transformer in a pipeline resulted in
only the last transformer being added to the pipeline's node tree.
Previous transformers were created but not persisted.

This uses the new Node.Find method to attach the node to the correct
portion of the tree below the root node.