Change default URI / DSN, strip prefix if there is one
Just `/` worked for the client test, but with the reader tests we'll
run into issues like this:

```
panic: default addr for network '127.0.0.1' unknown
```

Using 127.0.0.1 or `localhost` if we don't specify the protocol. And if
we are doing it for the reader tests we might as well do it for the
default.

---

Since everything else in the world assumes a prefix we can strip off
the prefix if one is supplied in the URI. Allows better consistency between Postgresql and MySQL