Update types for MySQL, insert data for reader tests
Getting closer. Worked through this list:

https://dev.mysql.com/doc/refman/5.7/en/data-types.html

And defined all the column types in order at the top of
`adaptor_test.go` and then in the same order at the bottom of the file.
It's now progressing further as doesn't complain about inserting data,
but is failing to read the row count for some reason - although data is
there.

That probably means `reader.go` needs altering if it's not pulling data.

Still need to update `reader_test.go` similarly.

Some types I've had to drop out for now as realised I'm only running 5.6
locally.