TestReadComplex - binary column passing
Figure something out for this. I'm honestly not sure how best to handle
this (and similar bit and spatial stuff).

I.e. should we actual do some decoding in reader? Or just do per this
commit and manipulate stuff in testing just so we can compare values?

Will probably come back and look at this again later on.