Reference commit - Trying to test Point values
Doesn't work, I get:

```
    reader_test.go:143: Expected colpoint of row to equal POINT(15 15) (string), but was [7.291122019556398e-304 2.5124430165029e-310] (orb.Point)
```

But commiting for reference as one way to try to use paulmach/orb. 
There are other options.

Maybe for Spatial stuff we do need to process in reader.go though and 
not in the test. That's going to take me a bit to think about though.