Reference commit - Working test for colpoint
It works and passes... but needs a lot of tidying up. Commiting before 
I change and accidentally break it (and then forget how I got it 
working).

One advantage of go-geom over orb is that is has wkbhex functions...

If we get the Hex string from MySQL then we get it exactly as [per this 
doc][1]. We can then use wkbhex to decode to go-geom geometry format 
and a further function to change it to wkt

Obviously there are various ways we could approach this. Instead of 
changing the value from MySQL to wkt we could leave it as go-geom and 
then make sure the value we are testing it against is also go-geom. At 
some point we probably want to store/reference spatial stuff internally 
using go-geom (or Orb - personally I like the look of Orb more, but 
go-geom is working) as that'll probably help with Postgresql (and 
transferring Spatial data between Postgresql and MySQL).

Made some minor spacing changes to suit how go-geom converts stuff to 
WKT.


[1]: https://dev.mysql.com/doc/refman/5.6/en/gis-data-formats.html