WIP - Wrap geometry in `ST_GeomFromText` at insert
Honestly... a 50 character limit on commit titles is shit. The above is 
a bit misleading. This is better:

Store geometry values in a Go geometry format so we can inspect their 
type in writer.go and wrap in `ST_GeomFromText` for the INSERT.

This is currently failing with:

```
--- FAIL: TestComplexInsert (0.02s)
    writer_test.go:171: unexpected Insert error, Error 1416: Cannot get geometry object from data you send to the GEOMETRY field
```

But it seems like it should work. I've verified a manual:

```
INSERT INTO writer_complex_insert_test.complex_test_table (...) VALUES (...actual values...)
```

using the text dumped out and all the `ST_GeomFromText` bits do work 
and do insert ok. So I'm not really sure where it's going wrong at the 
moment. Guess I need to dig into the underlying driver and understand 
what it's actually doing.