TestUpdate passes
We may want to think about using more strict sorting, but hopefully a 
later test (the next one, `TestComplexUpdate`) will confirm that if 
multiple primary keys are used.

But for now, since MySQL has more simple placeholders than Postgresql 
we need to manipulate the data more to get it in the correct order for 
the UPDATE query.

So we keep track of `cvals` and `uvals` separately so we can combine 
into one `vals` at the end in (hopefully) the correct order.

If we need to do it in a more detailed way we can build a struct so we 
can include `i` with the `key` and `value` and then do some sorting at 
the end.

Also... had to change the query for finding primary keys since it was 
still the Postgresql one. Stackoverflow to the rescue.