TestComplexUpdate passes
The main issue here was an "off by one" error in the test. The 
difference between the Postgresql and MySQL setup is that we are not 
using an integer counter in a loop to populate the ID, e.g:

```
for i := 0; i < data.InsertCount; i++ {
```

Instead we send `NULL` and let MySQL auto-increment. So in MySQL the ID 
is 1..10.

This is why in the test we've had to change from `ranInt` to `ranInt+1` 
in various places.

Another change was updating the `TestComplexUpdate` data per the 
`TestComplexInsert`, etc. I.e. it was still in Postgresql format.

And the final change was adjusting the placeholders in `updateMsg` per 
`insertMsg` to suit MySQL (`?` instead of `$1`, etc) and wrapping 
geometry placeholders in `ST_GeomFromText`.