TestComplexDelete passes
Noticed an off by one issue here as well. Same reason as under 
`TestComplexUpdate` so it's valid/ok. BUT the test was passing because 
of this error/issue! Since it was trying to delete a row that didn't 
existed it also checked for a row that didn't exist and so thought 
things were as expected. I added a row count check as well to cover 
this scenario.