Writer tests pass - handle scanning time values
```
unsupported Scan, storing driver.Value type []uint8 into type *time.Time
```

With the previous driver, 
[go-sql-driver/mysql](https://github.com/go-sql-driver/mysql), there 
was a handy 
[`parseTime`](https://github.com/go-sql-driver/mysql#parsetime) option 
that did this automagically. With 
[go-mysql-org/go-mysql](https://github.com/go-mysql-org/go-mysql) we 
need to do this manually - in the tests at least. With the actual 
adaptor we should be able to manage in the `reader.go` file (we already 
try to do that).