Disable Foreign Key Checks on write
Because otherwise it's not going to be possible to "import" data, etc. 
People will not always have access to disable globally on the sink.

This is just a rough/initial commit that I really hope I'll improve as 
it doesn't feel great sending this for every write.

However, I thought about querying the value and then setting only if 
needed, but that on it's own is not going to improve things, i.e. it's 
still a read for every write.

What's really needed is to hook this into the session initiation.