Reference commit for MySQL tailing
Just focused on fleshing out the `Read` function and making it look 
correct and changing everything else just enough to get it to build - 
the `processEvent` and `parseEvent` functions are _definitely_ not 
ready or correct yet and still have stuff that is specific to 
Postgresql replication slots.

The focus in `Read` has been:

1. Figuring out how to get hold of the DSN so that we can use that to 
build the configuration for the Sync client
2. Using the existing Transporter MySQL client ("self") to query for 
the binlog info
3. Understanding the select / case and channel setup

For the DSN I'd hoped we could do something like `client.dsn`, but in 
the end it's probably just easier to pass it through when the tailer is 
initiated. For Postgresql we pass through the replication slot, for 
MySQL we can just pass through the DSN.

Since this just builds and is untested I'm hoping obtaining the binlog 
info in this way will work:

```
result, err := session.mysqlSession.Query("SHOW MASTER STATUS")
result.Scan(&binFile, &binPosition)
```

I.e. using the Transporter client itself, not creating a new one, which 
means using the database/sql approach and scanning values in. It 
_should_ work, but if need be we could use the 
`go-mysql-org/go-mysql/client` package - see the example code in the 
README.

There are a lot of comments regarding the select/case and channel usage 
which is mostly me trying to understand how this was working for 
Postgresql and therefore what changes are needed for MySQL. Will try to 
summarise here as well:

For Postgresql we have a blocking 1 second sleep to give the `done` 
channel a chance to execute. Assuming it doesn't execute we make one 
call/query to logical decoding that pulls all changes since the last 
call. We then process this block of changes (multiple rows of changes).

For MySQL we have a streaming sync client instead and each call to the 
client fetches just one change/event. Blocking for 1 second therefore 
makes no sense. So to keep the same `done` channel approach I'm trying 
executing the MySQL sync stuff in the `default` section of the 
select/case. This way it won't block but should still allow `done` to 
execute. Possibly we can change this. We could also look at using 
`DumpEvents` instead of `GetEvent` which might be more similar to 
Postgresql, but I can't find much info on `DumpEvents`.