Review feedback - Avoid GetEvent blocking done
`GetEvent` is a blocking function:

https://github.com/go-mysql-org/go-mysql/blob/d1666538b005e996414063695ca223994e9dc19d/replication/binlogstreamer.go#L23-L25

Which means if no events are received it will still block the `done` 
channel as it'll just sit there and wait. Need to use a timeout context 
to work-around this so that if no events are received after 60 seconds 
(I think we can wait quite a while for Transporter's use case, but can 
change this later if needed) it will allow the done channel to process 
if anything is there. If not then it carries on with `GetEvent` again.

If there are events then things work as expected. I.e. it processes one 
event, loops and will check the done channel each time so we can break 
tailing faster if events are being sent.