Reference commit for MySQL tailing, processEvent
Just focused on `processEvent` for this commit and so what is in 
`parseEventData` is a load of crap. I just need enough changes there to 
get things to build.

As suspected we can copy/crib a lot from Canal and how that uses the 
replication package.

In Postgresql we get a long string from logical decoding which we then 
run a regex on to split up into:

1. Schema
2. Table
3. Action (Insert / Update / Delete)
4. Data

Also in Postgresql we have multiple rows of this to process that are 
looped through within `pluckFromLogicalDecoding`. We send just step 4 
(the Data) to `parseLogicalDecodingData`.

MySQL is a bit different. At the moment at least. We have just one 
event we deal with in `processEvent`. The Replication package also 
gives us a lot more info so rather than parse a long string we can use 
the Replication package to tell use the action on an event, etc.

The trickiest bit is that compared to Postgresql where we'll have a 
string like:

    "id[integer]:1 data[text]:'1'"

which we then parse with `parseLogicalDecodingData` to get into 
Transporter data format:

    map[data:1 id:1]

with MySQL we get a nested slice like this:

    [[1 Tacos] [2 Tomato Soup] [3 Grilled Cheese]]

with no column names or data types. This means processing this is going 
to be very different, but that's a problem for future me.