Reference commit - parseEventData builds
Doesn't work yet, but the logic is getting there. I think the main
thing I need to do is to move one of the loops back to `processEvent`.
Maybe.

With Postgresql we loop within `pluckFromLogicalDecoding` (the
`processEvent`) equivalent because we get can multiple lines from
logical decoding. I was thinking we get just one event from MySQL.
Which is true, but need to think about how we handle inserting multiple
rows like:

```
INSERT INTO recipes
    (recipe_id, recipe_name)
VALUES
    (1,"Tacos"),
    (2,"Tomato Soup"),
    (3,"Grilled Cheese");
```

We get this from the binlog:

```
[[1 Tacos] [2 Tomato Soup] [3 Grilled Cheese]]
```

Which is three lots of data. How do we convert that to the `data.Data`?
I.e. I'm just not sure if that needs to be three separate `data.Data`s
or whether it can be wrapped up as one?

But the basic principle is hopefully ok here: We don't need to worry
about `TableMapEvent`s we can just use Transporter itself and copy the
code from `iterateTable` in reader.go. Ideally we'd have a common
function that could be used by both, but for now copy/paste/adjust.

I think this _should_ work once I figure out how we deal with multiple
inserts. I might have to understand how Postgresql would behave.