Reference commit - loop through eventRows one level up
Another reference commit. Builds, probably doesn't run yet, is 
definitely inefficient, but basic logic should be correct.

This more closely resembles how Postgresql does things.

Previously I was under the impression that one binlog event was 
equivalent to one logicial decoding row, but that's not the case. One 
binlog event is actually equivalent to multiple logical decoding rows. 
I.e. the three insert rows here for Postgresql:

```
compose=# SELECT * FROM pg_logical_slot_get_changes('regression_slot', NULL, NULL);
    lsn    | xid |                                                              data                                                              
-----------+-----+--------------------------------------------------------------------------------------------------------------------------------
 0/500CEC8 | 496 | BEGIN 496
 0/500CEC8 | 496 | table public.recipes: INSERT: recipe_id[integer]:1 recipe_name[character varying]:'Tacos' recipe_rating[integer]:null
 0/500D050 | 496 | table public.recipes: INSERT: recipe_id[integer]:2 recipe_name[character varying]:'Tomato Soup' recipe_rating[integer]:null
 0/500D120 | 496 | table public.recipes: INSERT: recipe_id[integer]:3 recipe_name[character varying]:'Grilled Cheese' recipe_rating[integer]:null
 0/500D228 | 496 | COMMIT 496
(5 rows)
```

Are what we see here for this MySQL binlog event:

```
[[1 Tacos] [2 Tomato Soup] [3 Grilled Cheese]]
```

Therefore, like Postgresql does this in `pluckFromLogicalDecoding`:

```
for changesResult.Next() {
```

We need to do similarly in `processEvent`:

```
for _, row := range rowsEvent.Rows {
```

The main issue now is that we are being inefficient in `parseEventRow` 
because we are doing the columns and data types every single time.

This needs to be moved up a level too ideally. Or some of it at least. 
We only need to find columns and data types once per binlog event.

That'll be the next commit.