Move fetching column info up one level
As mentioned in the previous commit. Moves finding `column` names and 
data types up into the `processEvent` call so we do once per binlog 
event instead of doing it everytime in `parseEventRow` for each row of 
the binlog event. Much more efficient.

Could be made more efficient still in the future by storing that for 
all tables of the database.

The only downside of this is that some of `iterateTable` from reader.go 
is now split between `parseEventRow` and `processEvent` meaning we'd 
have to have two functions if we want to share between these locations. 
Still do-able though.

Loads of comments to clean-up and haven't started on the tailing tests 
yet, but in my mind this should work in principal. Bound to be little 
things I've missed and got wrong though.