Make WithCustomTLS pass through if no TLS options
I.e. how I had it before required TLS options to be specified, but we 
also want it to work if TLS options aren't specified. So we check for 
an empty `cert` string.

Also, the `c.db` bit has changed now, especially with changes to the 
driver using a `/` as a path separator. The `c.uri` already has a 
`mysql://` prefix/scheme.