Tidy up debug logging
As much as possible.

1. The Debug field in the adaptor config (copied from Postgresql) seems 
to be nonsense and doesn't do anything. Transporter has a top-level 
concept of debug logging. So remove that field/option
2. In a lot of places I had `fmt.Println(X)` commented out which I'd 
used for "debugging" and developing. I left a lot of TODOs to remove 
these. But instead of doing that we can try to make actual debug 
logging statements
3. Can't do that in the tests though so I have retained as commented 
out, but updated the comments to clarify why

I am sure logging could be improved further. Please have at it.