Remove `case []interface{}` from MySQL writer
After playing about and looking at this more, I _think_ that code (which
we copied from Postgresql) was for handling data like:

```
"colarrayint":        []interface{}{1, 2, 3, 4},
```

Which definitely isn't applicable to MySQL as it doesn't have array
support like Postgresql.

We can always revert this at a later point if it turns out we do need
it. But it doesn't stop the tests passing right now.